+++ src/lib/path/module.test.ts
@@ -1,46 +1,46 @@
 import {suite} from 'uvu';
 import * as assert from 'uvu/assert';
 
-import {isExternalModule} from './module.js';
+import {is_external_module} from './module.js';
 
-/* test__isExternalModule */
-const test__isExternalModule = suite('isExternalModule');
+/* test__is_external_module */
+const test__is_external_module = suite('is_external_module');
 
-test__isExternalModule('internal browser module patterns', () => {
-	assert.is(isExternalModule('./foo'), false);
-	assert.is(isExternalModule('./foo.js'), false);
-	assert.is(isExternalModule('../foo'), false);
-	assert.is(isExternalModule('../foo.js'), false);
-	assert.is(isExternalModule('../../../foo'), false);
-	assert.is(isExternalModule('../../../foo.js'), false);
-	assert.is(isExternalModule('/foo'), false);
-	assert.is(isExternalModule('/foo.js'), false);
-	assert.is(isExternalModule('src/foo'), false);
-	assert.is(isExternalModule('src/foo.js'), false);
-	assert.is(isExternalModule('$lib/foo'), false);
-	assert.is(isExternalModule('$lib/foo.js'), false);
-	assert.is(isExternalModule('./foo/bar/baz'), false);
-	assert.is(isExternalModule('./foo/bar/baz.js'), false);
-	assert.is(isExternalModule('../foo/bar/baz'), false);
-	assert.is(isExternalModule('../foo/bar/baz.js'), false);
-	assert.is(isExternalModule('../../../foo/bar/baz'), false);
-	assert.is(isExternalModule('../../../foo/bar/baz.js'), false);
-	assert.is(isExternalModule('/foo/bar/baz'), false);
-	assert.is(isExternalModule('/foo/bar/baz.js'), false);
-	assert.is(isExternalModule('src/foo/bar/baz'), false);
-	assert.is(isExternalModule('src/foo/bar/baz.js'), false);
-	assert.is(isExternalModule('$lib/foo/bar/baz'), false);
-	assert.is(isExternalModule('$lib/foo/bar/baz.js'), false);
+test__is_external_module('internal browser module patterns', () => {
+	assert.is(is_external_module('./foo'), false);
+	assert.is(is_external_module('./foo.js'), false);
+	assert.is(is_external_module('../foo'), false);
+	assert.is(is_external_module('../foo.js'), false);
+	assert.is(is_external_module('../../../foo'), false);
+	assert.is(is_external_module('../../../foo.js'), false);
+	assert.is(is_external_module('/foo'), false);
+	assert.is(is_external_module('/foo.js'), false);
+	assert.is(is_external_module('src/foo'), false);
+	assert.is(is_external_module('src/foo.js'), false);
+	assert.is(is_external_module('$lib/foo'), false);
+	assert.is(is_external_module('$lib/foo.js'), false);
+	assert.is(is_external_module('./foo/bar/baz'), false);
+	assert.is(is_external_module('./foo/bar/baz.js'), false);
+	assert.is(is_external_module('../foo/bar/baz'), false);
+	assert.is(is_external_module('../foo/bar/baz.js'), false);
+	assert.is(is_external_module('../../../foo/bar/baz'), false);
+	assert.is(is_external_module('../../../foo/bar/baz.js'), false);
+	assert.is(is_external_module('/foo/bar/baz'), false);
+	assert.is(is_external_module('/foo/bar/baz.js'), false);
+	assert.is(is_external_module('src/foo/bar/baz'), false);
+	assert.is(is_external_module('src/foo/bar/baz.js'), false);
+	assert.is(is_external_module('$lib/foo/bar/baz'), false);
+	assert.is(is_external_module('$lib/foo/bar/baz.js'), false);
 });
 
-test__isExternalModule('external browser module patterns', () => {
-	assert.is(isExternalModule('foo'), true);
-	assert.is(isExternalModule('foo.js'), true);
-	assert.is(isExternalModule('foo/bar/baz'), true);
-	assert.is(isExternalModule('foo/bar/baz.js'), true);
-	assert.is(isExternalModule('@foo/bar/baz'), true);
-	assert.is(isExternalModule('@foo/bar/baz.js'), true);
+test__is_external_module('external browser module patterns', () => {
+	assert.is(is_external_module('foo'), true);
+	assert.is(is_external_module('foo.js'), true);
+	assert.is(is_external_module('foo/bar/baz'), true);
+	assert.is(is_external_module('foo/bar/baz.js'), true);
+	assert.is(is_external_module('@foo/bar/baz'), true);
+	assert.is(is_external_module('@foo/bar/baz.js'), true);
 });
 
-test__isExternalModule.run();
-/* test__isExternalModule */
+test__is_external_module.run();
+/* test__is_external_module */