+++ src/lib/path/module.ts
@@ -10,5 +10,5 @@ const INTERNAL_MODULE_MATCHER = new RegExp(
 	'u',
 );
 
-export const isExternalModule = (moduleName: string): boolean =>
+export const is_external_module = (moduleName: string): boolean =>
 	!INTERNAL_MODULE_MATCHER.test(moduleName);